# specs/openapi.yaml
openapi: 3.0.0
info:
  title: Study Buddy
  version: 1.0.0
  description: Study Budy Api
servers:
  - url: https://localhost:3000
    description: Local development server

  - url: https://studdy-buddy-api-h7kw3.ondigitalocean.app/
    description: Digital Ocean Server
paths:
...
  /group/invite/{chatId}/{userId}:
    post:
      summary: Invite User to Group
      operationId: postGroupInvite
      tags:
        - Groups
      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: The receiver ID
        - in: path
          name: chatId
          required: true
          schema:
            type: string
          description: chatId of Group
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
  /group/invite/{inviteId}:
    patch:
      summary: Handle Group Invite
      operationId: handleGroupInvite
      tags:
        - Groups
      security:
        - bearerAuth:
            - read:users
            - public
      parameters:
        - in: path
          name: inviteId
          required: true
          schema:
            type: string
          description: The request ID

      requestBody:
        content:
          application/json:
            schema:
              properties:
                isAccepted:
                  type: boolean

              required:
                - isAccepted
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"

  #   delete:
  #     summary: Delete Friend Request
  #     operationId: deleterequest
  #     tags:
  #       - Friends
  #     security:
  #       - bearerAuth:
  #           - read:users
  #           - public
  #     parameters:
  #       - in: path
  #         name: requestId
  #         required: true
  #         schema:
  #           type: string
  #         description: The request ID
  #     responses:
  #       "200":
  #         description: Success
  #       "400":
  #         description: Bad request
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/BadRequestError"
  /group/invites:
    get:
      summary: Get Group Invites
      operationId: getGroupInvites
      tags:
        - Groups
      security:
        - bearerAuth:
            - read:users
            - public

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Request"

        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
  /groups:
    get:
      summary: Get all Groups of User
      operationId: getAllUserGroups
      tags:
        - Groups

      security:
        - bearerAuth:
            - read:users
            - public

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"

        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"

  # /group/{chatId}/{userId}:
  #   delete:
  #     summary: Remove User from group
  #     operationId: removeFriend
  #     tags:
  #       - Friends

  #     security:
  #       - bearerAuth:
  #           - read:users
  #           - public

  #     parameters:
  #       - in: path
  #         name: friendId
  #         required: true
  #         schema:
  #           type: string
  #         description: The ID of friend being removed

  #     responses:
  #       "200":
  #         description: Success
  #       "400":
  #         description: Bad request
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/BadRequestError"
  /group/{chatId}:
    get:
      summary: Get group
      operationId: getGroup
      tags:
        - Groups

      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
          description: The chatId of Group

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                properties:
                  owner:
                    type: string
                    format: objectId
                  groupName:
                    type: string
                  users:
                    type: array
                    items:
                      type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
    delete:
      summary: Delete group
      operationId: deleteGroup
      tags:
        - Groups

      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
          description: The ID of Group being deleted

      responses:
        "200":
          description: Success
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
...
  /group/{chatId}/message:
    post:
      summary: Send a message to a group chat
      operationId: sendMessageToGroup
      tags:
        - Messages
      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: chatId
          required: true
          description: ID of the group chat
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: "hello"
              required:
                - content
      responses:
        "200":
          description: Message successfully sent to the group
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                    example: "hello"
                  sender:
                    type: string
                    description: ID of the user who sent the message
        "400":
          description: Chat does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /group/{chatId}/messages:
    get:
      summary: Get messages from a group chat
      operationId: getMessagesFromGroup
      tags:
        - Messages
      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: chatId
          required: true
          description: ID of the group chat to retrieve messages from
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved group messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    sender:
                      type: string
                    content:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
        "400":
          description: Chat does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized
        "500":
          description: Server error
components:
  schemas:
    User:
      type: object
      required:
        - firstName
        - lastName
        - username
        - email
        - password
      properties:
        firstName:
          type: string

        lastName:
          type: string

        username:
          type: string
          minLength: 4

        email:
          type: string

        password:
          type: string
          minLength: 8

        friends:
          type: array
          items:
            $ref: "#/components/schemas/Friend"

        incomingRequests:
          type: array
          items:
            $ref: "#/components/schemas/Request"

        outgoingRequests:
          type: array
          items:
            $ref: "#/components/schemas/Request"

        chatSessions:
          type: array
          items:
            $ref: "#/components/schemas/Chat"

      example:
        firstName: "Cornelius"
        lastName: "Bartels"
        email: "cjkb@gmail.com"
        password: "test121111"
    BadRequestError:
      type: object
      properties:
        name:
          type: string
        errors:
          type: array
          items:
            type: object
...
    Request:
      type: object
      required:
        - sender
        - receiver
      properties:
        sender:
          $ref: "#/components/schemas/User"
        receiver:
          $ref: "#/components/schemas/User"
        isAccepted:
          type: boolean
    Friend:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    Chat:
      type: object
      properties:
        chatType:
          type: string
          enum:
            - direct
            - group
        messageBuckets:
          type: array
          items:
            $ref: "#/components/schemas/User"
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        groupName:
          type: string
        owner:
          $ref: "#/components/schemas/User"

      required:
        - chatType
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt